# General improvement of the TextLogging component