In details:
- If the project is compiled in `debug` the spd level is set to debug
- Use Meyers' implementation to instantiate the singleton. We do not need to use `std::once_flag`
- Avoid to use `spdlog::get()` continuously as suggested in the [`spdlog` documentation](https://github.com/gabime/spdlog/wiki/2.-Creating-loggers#accessing-loggers-using-spdlogget)   

I also run the following code to benchmark the new implementation and the old one.

**Disclaimer: I don't see huge difference**

```cpp
#include <chrono>
#include <thread>

#include <BipedalLocomotion/TextLogging/Logger.h>

using namespace std;

void foo() {
    auto start = chrono::steady_clock::now();
    BipedalLocomotion::log()->warn("Hello");
    auto end = chrono::steady_clock::now();
    BipedalLocomotion::log()->info("Hello - Time: {} ms", chrono::duration<double>(end - start).count() * 1e3);

    start = chrono::steady_clock::now();
    BipedalLocomotion::log()->warn("I'm still here!");
    end = chrono::steady_clock::now();
    BipedalLocomotion::log()->info("I'm still here! - Time: {} ms", chrono::duration<double>(end - start).count() * 1e3);
}

int main(int argc, char **argv) {
    std::vector<std::thread> threadVector;

    for(int i = 0; i < 5; i++)
        threadVector.emplace_back(foo);

    for(auto& t: threadVector)
        t.join();

    return 0;
};
```

### New Implementation
```
[2021-03-31 11:24:07.953] [thread: 34018] [blf] [warning] Hello
[2021-03-31 11:24:07.954] [thread: 34018] [blf] [info] Hello - Time: 0.349754 ms
[2021-03-31 11:24:07.954] [thread: 34018] [blf] [warning] I'm still here!
[2021-03-31 11:24:07.954] [thread: 34018] [blf] [info] I'm still here! - Time: 0.016776 ms
[2021-03-31 11:24:07.953] [thread: 34022] [blf] [warning] Hello
[2021-03-31 11:24:07.954] [thread: 34022] [blf] [info] Hello - Time: 0.313552 ms
[2021-03-31 11:24:07.953] [thread: 34019] [blf] [warning] Hello
[2021-03-31 11:24:07.954] [thread: 34019] [blf] [info] Hello - Time: 0.57089 ms
[2021-03-31 11:24:07.954] [thread: 34019] [blf] [warning] I'm still here!
[2021-03-31 11:24:07.954] [thread: 34019] [blf] [info] I'm still here! - Time: 0.012905 ms
[2021-03-31 11:24:07.953] [thread: 34021] [blf] [warning] Hello
[2021-03-31 11:24:07.954] [thread: 34021] [blf] [info] Hello - Time: 0.57242 ms
[2021-03-31 11:24:07.954] [thread: 34021] [blf] [warning] I'm still here!
[2021-03-31 11:24:07.954] [thread: 34022] [blf] [warning] I'm still here!
[2021-03-31 11:24:07.954] [thread: 34022] [blf] [info] I'm still here! - Time: 0.25409800000000005 ms
[2021-03-31 11:24:07.953] [thread: 34020] [blf] [warning] Hello
[2021-03-31 11:24:07.954] [thread: 34020] [blf] [info] Hello - Time: 0.816744 ms
[2021-03-31 11:24:07.954] [thread: 34020] [blf] [warning] I'm still here!
[2021-03-31 11:24:07.954] [thread: 34020] [blf] [info] I'm still here! - Time: 0.007591 ms
[2021-03-31 11:24:07.954] [thread: 34021] [blf] [info] I'm still here! - Time: 0.007428 ms

```

### Old implementation
```
[2021-03-31 11:27:35.587] [thread: 35013] [blf] [warning] Hello
[2021-03-31 11:27:35.587] [thread: 35009] [blf] [warning] Hello
[2021-03-31 11:27:35.587] [thread: 35009] [blf] [info] Hello - Time: 0.49555400000000005 ms
[2021-03-31 11:27:35.587] [thread: 35009] [blf] [warning] I'm still here!
[2021-03-31 11:27:35.587] [thread: 35009] [blf] [info] I'm still here! - Time: 0.02586 ms
[2021-03-31 11:27:35.587] [thread: 35013] [blf] [info] Hello - Time: 0.202951 ms
[2021-03-31 11:27:35.587] [thread: 35013] [blf] [warning] I'm still here!
[2021-03-31 11:27:35.587] [thread: 35012] [blf] [warning] Hello
[2021-03-31 11:27:35.587] [thread: 35011] [blf] [warning] Hello
[2021-03-31 11:27:35.587] [thread: 35010] [blf] [warning] Hello
[2021-03-31 11:27:35.588] [thread: 35010] [blf] [info] Hello - Time: 0.903108 ms
[2021-03-31 11:27:35.588] [thread: 35010] [blf] [warning] I'm still here!
[2021-03-31 11:27:35.588] [thread: 35010] [blf] [info] I'm still here! - Time: 0.046075000000000005 ms
[2021-03-31 11:27:35.587] [thread: 35013] [blf] [info] I'm still here! - Time: 0.036882 ms
[2021-03-31 11:27:35.588] [thread: 35012] [blf] [info] Hello - Time: 0.675508 ms
[2021-03-31 11:27:35.588] [thread: 35011] [blf] [info] Hello - Time: 0.78243 ms
[2021-03-31 11:27:35.588] [thread: 35011] [blf] [warning] I'm still here!
[2021-03-31 11:27:35.588] [thread: 35012] [blf] [warning] I'm still here!
[2021-03-31 11:27:35.588] [thread: 35011] [blf] [info] I'm still here! - Time: 0.043532999999999995 ms
[2021-03-31 11:27:35.588] [thread: 35012] [blf] [info] I'm still here! - Time: 0.144198 ms
```

This closes #265