@GiulioRomualdi 2021-03-31
Windows CI Error
```
Chocolatey v0.10.15
Installing the following packages:
wget;unzip
By installing you accept licenses for the packages.
wget not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable
wget package files install completed. Performing other installation steps.
The install of wget was NOT successful.
wget not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable
unzip not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable
unzip package files install completed. Performing other installation steps.
The install of unzip was NOT successful.
unzip not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable

Chocolatey installed 0/2 packages. 2 packages failed.
 See the log for details (C:\ProgramData\chocolatey\logs\chocolatey.log).

Failures
 - unzip (exited 1) - unzip not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable
 - wget (exited 1) - wget not installed. An error occurred during installation:
 The remote server returned an error: (503) Server Unavailable. Service Unavailable
wget: D:\a\_temp\061fcd8a-518d-4089-9941-7cbe3a7d4300.ps1:7
Line |
   7 |  wget https://github.com/robotology/robotology-superbuild-dependencies …
     |  ~~~~
     | The term 'wget' is not recognized as a name of a cmdlet, function, script file, or executable program.
     | Check the spelling of the name, or if a path was included, verify that the path is correct and try
     | again.
```