@traversaro 2021-03-31
> 
> 
> Network flake I guess, if the conda based Windows CI we are good to go I guess.

I restarted the jobs in any case.